% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{calculateCost}
\alias{calculateCost}
\title{Compute the network travel cost matrix}
\usage{
calculateCost(graph, rasterName, fromName, fromField, toName, toField)
}
\arguments{
\item{graph}{graph object generated by \emph{createGraph()}}

\item{rasterName}{filename (character) of the raster representation of the
road network}

\item{fromName}{filename (character) of a point vector file with travel start
locations}

\item{fromField}{character, name of column in \emph{fromName} used as point
identifier}

\item{toName}{filename (character) of a point vector file with travel end
locations}

\item{toField}{character, name of column in \emph{toName} used as point
identifier}
}
\value{
data frame. The first column named \emph{fromField} contains the IDs
of the destination points. Each of the remaining columns is named by the
values in \emph{fromField} and reports the pairwise travel cost in hours
across \emph{fromName} and \emph{toName} points. An NA value corresponding to
\emph{fromName} to \emph{toName} pair indicates that there is no path along
the road network that connects them and suggests that one or both points in
the pair have not moved to the interconnected road network by using
movePt2RdSegment()
}
\description{
Calculates the travel cost, expressed in hours, between points
in \emph{fromName} and \emph{toName} respectively
}
\details{
\emph{fromName} and \emph{toName} locations should be on, or have
been moved to the network by applying \emph{movePt2RdSegement()}
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "data", "sample.gpkg", package="tRee2MillCost" )
toPath <- system.file( "data", "mills.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- createGraph( "road_net.tif" )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
names( terra::vect( "movedFromLocations.gpkg" ) )
movePt2RdSegment( "road_net.tif", toPath, "movedToLocations.gpkg" )
names( terra::vect( "movedToLocations.gpkg" ) )
cost.df <- calculateCost( graph, "road_net.tif",
                          "movedFromLocations.gpkg", "sampleID",
                          "movedToLocations.gpkg", "COMPANY_NAME" )
write.csv( cost.df, "cost_matrix.csv", row.names=FALSE )
}

}
