% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{connectivity}
\alias{connectivity}
\title{Determine if points are on value cells of a raster}
\usage{
connectivity(rasterName, pName)
}
\arguments{
\item{rasterName}{filename (character) of the network raster}

\item{pName}{filename (character) of a point vector}
}
\value{
A console message with connectivity results
}
\description{
Determine if points are on value cells of a raster
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "data", "sample.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- createGraph( "road_net.tif" )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
pOriginal <- terra::vect( fromPath )
pMoved <- terra::vect( "movedFromLocations.gpkg" )
connectivity( "road_net.tif", "movedFromLocations.gpkg" )
connectivity( "road_net.tif", fromPath )
}

}
