% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{costSummary}
\alias{costSummary}
\title{Summarize travel cost}
\usage{
costSummary(
  costDF,
  fromName,
  fromField,
  maxMoveDistance = "",
  travelThreshold = ""
)
}
\arguments{
\item{costDF}{data frame created by \emph{calculateCost}}

\item{fromName}{character, filename/path of the point vector layer
representing travel start location(s)}

\item{fromField}{character, the field in \emph{fromName} used with
\emph{calculateCost()} and representing travel start locations}

\item{maxMoveDistance}{optional numeric, the maximum 2D distance a
\emph{fromName} location is allowed to move to be on the road network and
still considered in the summary}

\item{travelThreshold}{optional numeric, maximum travel time, in hours, or
travel distance, in kilometers, to destination location(s) considered in the
summary}
}
\value{
list with two elements; a data frame comprising 'from' locations IDs,
the closest, travel time or distance to 'to' location IDs, and the
corresponding travel time, in hours, or distance, in kilometers. The second
list element, also a data frame, with 'to' location IDs, the number of 'from'
locations for which the 'to' location is closest in travel time or distance,
and the mean and standard deviation for the travel cost (time) for those
'from' locations
}
\description{
Processes a cost matrix to generate a summary
}
\details{
\emph{maxMoveDistance} should be in the linear unit of the
projection of the spatial data used to calculate the cost matrix. If both
\emph{maxMoveDistance} and \emph{travelThreshold} are missing or set to "",
all locations in \emph{fromName} will be used to create the summary.
\emph{maxMoveDistance} is often referred to as 'yarding' distance. Smaller
values for \emph{maxMoveDistance} and \emph{travelThreshold} result in fewer
'from' and 'to' locations included in the summary

\emph{fromName} must represent locations moved to the rasterized road network
by \emph{movePt2RdSegment()}
}
\examples{
\dontrun{
rdPath <- system.file( "extdata", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "extdata", "sample.gpkg", package="tRee2MillCost" )
toPath <- system.file( "extdata", "mills.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, 20, "road_net.tif", "MPH" )
graph <- createGraph( "road_net.tif", 1 )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
movePt2RdSegment( "road_net.tif", toPath, "movedToLocations.gpkg" )
cost.df <- calculateCost( graph, "road_net.tif",
                          "movedFromLocations.gpkg", "sampleID",
                          "movedToLocations.gpkg", "COMPANY_NAME" )
summary.list <- costSummary( cost.df,
                             fromName="movedFromLocations.gpkg",
                             fromField="sampleID",
                             travelThreshold=0.5 ) #30 minutes
print( summary.list[[1]] )
print( summary.list[[2]] )
}

}
\seealso{
\link{calculateCost} and \link{movePt2RdSegment}
}
