% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{createGraph}
\alias{createGraph}
\title{Create a routing graph from a road network raster}
\usage{
createGraph(rasterName, costType)
}
\arguments{
\item{rasterName}{character, filename/path of the road network raster}

\item{costType}{integer, 1 for travel calculations based on time, 2 for
travel calculations based on distance}
}
\description{
Relies on a raster representation of a road network
\emph{(rasterName)} and embedded functions to create a network graph. Network
cells with all 8 immediate neighbors representing background (no road) are
labeled orphan, saved into a file named orphan_xxxx.gpkg, and removed from
\emph{rasterName}. Segments with more than one cell each that are
disconnected from the rest of the network are saved into a file named
dc_xxxx.pgkg and are also removed from \emph{rasterName}
}
\details{
The frequency of orphan cells or disconnected cell groups (road
network segments) tends to increase as the resolution of the raster becomes
finer.
}
\examples{
\dontrun{
rdPath <- system.file( "extdata", "roads.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, 20, "road_net.tif", "MPH" )
graph <- createGraph( "road_net.tif", 1 )
}

}
\seealso{
\link[cppRouting:makegraph]{cppRouting::makegraph} and \link{prepareGraph}
}
