% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{maskCells}
\alias{maskCells}
\title{Set the value of selected raster cells to NA}
\usage{
maskCells(rasterName, IDs, tmpName)
}
\arguments{
\item{rasterName}{filename (character) of the raster}

\item{IDs}{numeric, vector of cell IDs}

\item{tmpName}{filename (character) filename of the output raster}
}
\description{
Set the value of selected raster cells to NA
}
\details{
Deprecated by the far more efficient \emph{terra::update()}
function. Will
be removed in future updates
}
