% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{prepareGraph}
\alias{prepareGraph}
\title{Create cell connectivity data frame from raster}
\usage{
prepareGraph(rasterName)
}
\arguments{
\item{rasterName}{filename (character) of a road network raster}
}
\value{
data frame with three columns. The first two are the IDs of cell
pairs, as inherited from the raster. The third column is the cost, in
seconds, to travel between the cell centers
}
\description{
Uses a raster representing a road network to create a data frame
featuring adjacent, and therefore connected, pairs of cells and calculates
the transition cost for each pair
}
\details{
The raster is expected to represent the road network speed limit, in
miles per hour. Although integers, the cell IDs in the output are of
numeric type, to prevent numerical overflow issues with rasters comprising
more than 2^31-1 cells. The output of this function is subsequently used to
create a graph on which all travel cost (time) computations are performed
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
df <- prepareGraph( "road_net.tif" )
head( df )
}

}
