% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{rasterizeRoads}
\alias{rasterizeRoads}
\title{Rasterize a line vector file}
\usage{
rasterizeRoads(shpName, shpField, resolution, rasterName)
}
\arguments{
\item{shpName}{filename (character) of a line vector}

\item{shpField}{name (character) of column expected to be present in
\emph{shpName}}

\item{resolution}{numeric, the resolution of the raster to be created}

\item{rasterName}{filename (character) of the raster to be created}
}
\description{
Rasterize a line vector file
}
\details{
\emph{shpName} is expected to a vector (line) representation of the
road network. Cells of the output raster overlapping more than one line
segments inherit the highest \emph{shpField} value. \emph{shpField} must
contain the speed limit for every line segment expressed in miles per hour
and in the (0, 199] range. \emph{resolution} coarser than 50m (164ft) will
induce unwarranted over-generalization of the road network. Very fine
\emph{resolution}, finer than 5m, can lead to a very large raster and routing
graph and likely yield numerous disconnected segments. If \emph{rasterName}
exists, it will be overwritten. TIFF format is expected. Datatype will be set
to INT1U with NODATA cell values set to 255
}
\examples{
\dontrun{
rasterizeRoads("roads.gpkg", "MPH", 20, "road_net.tif")
}

}
\seealso{
\link{checkInputs}
}
