% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{backgroundCellValue}
\alias{backgroundCellValue}
\title{Set raster background cell values to NA}
\usage{
backgroundCellValue(rasterName)
}
\arguments{
\item{rasterName}{filename (character) of the raster to be evaluated}
}
\description{
Function useful in fully automated processing of numerous road
network rasters contributed by third parties to determine their background,
no road, value. Background values of such rasters have been found to be 0,
<0, -9999, or 255. These values are incompatible with processing provisions
of this package
}
\details{
A cell value is labeled 'background' in \emph{rasterName} and is
subsequently set to NA if it represents at least 85\% for the raster. Road
network rasters are sparse matrices where cells corresponding to the road
network are collectively a small percentage of the total number of cells. If
the background value must change to NA, the value cells are encoded as
integers
}
