% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{calculateCost}
\alias{calculateCost}
\title{Compute the network cost matrix}
\usage{
calculateCost(graph, rasterName, fromName, fromField, toName, toField)
}
\arguments{
\item{graph}{graph object generated by \emph{createGraph()}}

\item{rasterName}{character, filename/path of the raster network}

\item{fromName}{character, filename/path of a point vector layer with travel
start locations}

\item{fromField}{character, name of column in \emph{fromName} used as point
identifier}

\item{toName}{character, filename/path of a point vector layer with travel
end locations}

\item{toField}{character, name of column in \emph{toName} used as point
identifier}
}
\value{
data frame. The first column named \emph{toField} contains the IDs
of the destination points. Each of the remaining columns is named by the
values in \emph{fromField} and reports the pairwise cost either as travel
time in hours or travel distance in kilometers across \emph{fromName} and
\emph{toName} points. An NA value indicates no path exists between the
corresponding \emph{fromName} and \emph{toName} locations and suggests that
one or both points in the pair have not moved using \emph{movePt2RdSegment()}
}
\description{
Calculates the travel cost, either in hours or kilometers,
between points in \emph{fromName} and \emph{toName} respectively.
}
\details{
\emph{fromName} and \emph{toName} locations should be on, or have
been moved to the network by applying \emph{movePt2RdSegement()}. The
\emph{costType} applied is read from the \emph{graph} object
}
\examples{
\dontrun{
rdPath <- system.file( "extdata", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "extdata", "sample.gpkg", package="tRee2MillCost" )
toPath <- system.file( "extdata", "mills.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, 20, "road_net.tif", "MPH" )
graph <- createGraph( "road_net.tif", 1 )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
names( terra::vect( "movedFromLocations.gpkg" ) )
movePt2RdSegment( "road_net.tif", toPath, "movedToLocations.gpkg" )
names( terra::vect( "movedToLocations.gpkg" ) )
cost.df <- calculateCost( graph, "road_net.tif",
                          "movedFromLocations.gpkg", "sampleID",
                          "movedToLocations.gpkg", "COMPANY_NAME" )
write.csv( cost.df, "cost_matrix.csv", row.names=FALSE )
}

}
