% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{cellFromGraphNode}
\alias{cellFromGraphNode}
\title{Get the road network cellID from a graph node}
\usage{
cellFromGraphNode(graph, node)
}
\arguments{
\item{graph}{object generated by \emph{createGraph()}}

\item{node}{numeric, the ID of a graph node}
}
\value{
numeric, the ID of a cell of the road network raster used to create
the graph. Or NA if the node does not exist or the graph
}
\description{
Determines the cellID of the road network raster from a node on
the graph created from the raster. The node must exist
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
graphPath <- system.file( "data", "graph_20m.rds", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- readRDS( graphPath )
node <- 18692
cellID <- cellFromGraphNode( graph, node )
node.xy <- terra::xyFromCell( terra::rast( "road_net.tif" ), cellID )
cat( paste0("Graph node=", node, ", cellID=", cellID, ", X=", node.xy[1], ", Y =", node.xy[2], "\n") )
}

}
\seealso{
\link{graphNodeFromXY} and \link{getIsochrone}
}
