% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{connectivity}
\alias{connectivity}
\title{Determine if points are on value cells of a raster}
\usage{
connectivity(rasterName, pName)
}
\arguments{
\item{rasterName}{character, filename/path of the network raster}

\item{pName}{character, filename/path of a point vector layer}
}
\value{
A console message with connectivity results
}
\description{
Determine if points are on value cells of a raster
}
\examples{
\dontrun{
rdPath <- system.file( "extdata", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "extdata", "sample.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, 20, "road_net.tif", "MPH" )
graph <- createGraph( "road_net.tif", 1 )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
pOriginal <- terra::vect( fromPath )
pMoved <- terra::vect( "movedFromLocations.gpkg" )
connectivity( "road_net.tif", "movedFromLocations.gpkg" )
connectivity( "road_net.tif", fromPath )
}

}
