% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{costSummary}
\alias{costSummary}
\title{Summarize travel cost}
\usage{
costSummary(
  costName,
  fromField,
  maxMoveDistance = NULL,
  maxTime = NULL,
  fromName = NULL
)
}
\arguments{
\item{costName}{filename (character) of the CSV generated by
\emph{calculateCost()}}

\item{fromField}{character, the field in \emph{fromName} used with
\emph{calculateCost()} and representing travel start locations}

\item{maxMoveDistance}{NULL (default) or a positive numeric value for the
maximum 2D distance a \emph{fromName} location is allowed to move to be on
the road network and still considered in the summary}

\item{maxTime}{NULL (default) or a positive numeric value for maximum travel
time, in hours, to destination location(s) considered in the summary}

\item{fromName}{NULL or filename (character) of the point vector representing
travel start location(s)}
}
\value{
list with two elements; a dataframe comprising 'from' locations IDs,
the closest, travel time-wise to 'to' location IDs, and the corresponding
travel time, in hours. The second element, also a data frame, with 'to'
location IDs, the number of 'from' locations for which the 'to' location is
closest, and the mean and standard deviation for the travel cost (time) for
those 'from' locations
}
\description{
Processes cost matrix to generate a summary
}
\details{
If \emph{maxMoveDistance} is not NULL, \emph{fromName} must be
specified. \emph{maxMoveDistance} should be in the unit of the projection of
the spatial data used to calculate the cost matrix. If both
\emph{maxMoveDistance} and \emph{maxTime} are set to NULL, all 'from'
locations in \emph{fromName} will be used to create the summary.
\emph{maxMoveDistance} is often referred to as 'yarding' distance. Smaller
values for \emph{maxMoveDistance} and \emph{maxTime} result in fewer
'from' and 'to' locations included in the summary

\emph{fromName} must represent locations moved to the rasterized road network
by \emph{movePt2RdSegment()}
}
\examples{
\dontrun{
costDataPath <- system.file( "data", "cost_matrix.csv", package="tRee2MillCost" )
summary.list <- costSummary( costDataPath, fromField="plotID", maxTime=0.5)
print( summary.list[[1]] )
print( summary.list[[2]] )
}

}
\seealso{
\link{calculateCost} and \link{movePt2RdSegment}
}
