% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{createGraph}
\alias{createGraph}
\title{Create a routing graph from a road network raster}
\usage{
createGraph(rasterName)
}
\arguments{
\item{rasterName}{filename (character) of the road network raster}
}
\description{
Relies on a raster representation of a road network
\emph{(rasterName)} and embedded functions to create a network graph. Network
cells all 8 immediate neighbors representing background (no road) are labeled
orphan, are saved into a file named orphan_xxxx.gpkg, and are removed from
\emph{rasterName}. Segments with more than one cell each that are
disconnected from the rest of the network are saved into a file named
dc_xxxx.pgkg and are also removed from \emph{rasterName}
}
\details{
The frequency of orphan cells or disconnected cell groups (road
network segments) tends to increase as the resolution of the raster becomes
finer.
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- createGraph( "road_net.tif" )
}

}
\seealso{
\link[cppRouting:makegraph]{cppRouting::makegraph} and \link{prepareGraph}
}
