% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{fromToPaths}
\alias{fromToPaths}
\title{delineate routes between two sets of locations}
\usage{
fromToPaths(
  graph,
  roadRasterName,
  fromName,
  fromField,
  fromIDs,
  toName,
  toField,
  toIDs
)
}
\arguments{
\item{graph}{object generated by \emph{createGraph()}}

\item{roadRasterName}{filename (character) of the road raster}

\item{fromName}{filename (character) of a point vector file}

\item{fromField}{character, the name of a field with point IDs present in
\emph{fromName}}

\item{fromIDs}{character, vector of selected point IDs, present in
\emph{fromField}, representing path origins}

\item{toName}{filename (character) of a point vector file}

\item{toField}{character, the name of a field with point IDs present in
\emph{toName}}

\item{toIDs}{character, vector of selected point IDs, present in
\emph{toField}, representing path destinations}
}
\value{
spatVect line object
}
\description{
Uses a road raster and a graph created from the raster to
delineate the routes between two sets of locations present in the graph
}
\details{
Originally designed to identify the routes from a collection of
points to a single destination point, as from sample locations (forest
inventory plots) to a wood processing facility. It can also be used with a
set of destination points. The routes identified carry no attributes

\emph{fromIDs} and \emph{toIDs} for locations of interest can be determined
by querying the road raster or the graph. All IDs must be present on the
graph. Using a contracted or simplified graph requires that all IDs have been
kept in \emph{cppRouting::cpp_simplify()} and even then the paths delineated
will likely deviate from the original vector road representation used to
great the graph. To maintain path detail, a non-optimized version of the
graph should be used at the cost of longer processing time
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
graphPath <- system.file( "data", "graph_20m.rds", package="tRee2MillCost" )
toPath <- system.file( "data", "mills.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "data", "sample.gpkg", package="tRee2MillCost" )

rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- readRDS( graphPath )
movePt2RdSegment( "road_net.tif", toPath, "movedToLocations.gpkg" )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )

r <- terra::rast( "road_net.tif" )

from.v <- terra::vect( "movedFromLocations.gpkg" )
#randomly select 5 'from' points
from.v <- from.v[sample( 1:length(from.v), 5, replace=FALSE )]
fromIDs <- from.v$sampleID

to.v <- terra::vect( "movedToLocations.gpkg" )
#randomly select 1 'to' point
to.v <- to.v[sample( 1:length(to.v), 1 )]
toIDs <- to.v$COMPANY_NAME
paths <- fromToPaths( graph, "road_net.tif",
                      "movedFromLocations.gpkg", "sampleID", fromIDs,
                      "movedToLocations.gpkg", "COMPANY_NAME", toIDs )
terra::plot( paths )
terra::points( from.v, pch=16, col="red", cex=1.5 )
terra::points( to.v, pch=16, col="blue", cex=1.5 )
}

}
\seealso{
\link{costSummary}, \link[cppRouting:cpp_simplify]{cppRouting::cpp_simplify}, and \link{calculateCost}
}
