% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{graphNodeFromXY}
\alias{graphNodeFromXY}
\title{Get the graph node ID using coordinates}
\usage{
graphNodeFromXY(graph, rasterName, X, Y)
}
\arguments{
\item{graph}{object generated by \emph{createGraph()}}

\item{rasterName}{character, filename/path of the raster network used to
create the graph}

\item{X}{Numeric, X coordinate}

\item{Y}{Numeric, Y coordinate}
}
\description{
Returns the graph node ID at specified coordinates
Graph note IDs shall not be confused with the cell IDs of the road network
raster used to create the graph
}
\examples{
\dontrun{
rdPath <- system.file( "extdata", "roads.gpkg", package="tRee2MillCost" )
graphPath <- system.file( "extdata", "graph_20m_time.rds", package="tRee2MillCost" )
rasterizeRoads( rdPath, 20, "road_net.tif", "MPH" )
# Since rasterizeRoads() has the rdField populated the graph loaded should be
# based on travel time
graph <- readRDS( graphPath )
x <- 496350
y <- 4903210
graphNodeID <- graphNodeFromXY( graph, terra::rast("road_net.tif"), X=x, Y=y )
cat( graphNodeID )
y <- 4903230
graphNodeID <- graphNodeFromXY( graph, terra::rast("road_net.tif"), X=x, Y=y )
}

}
