% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{movePt2RdSegment}
\alias{movePt2RdSegment}
\title{Move points in a vector file to road network cells}
\usage{
movePt2RdSegment(rasterName, inName, outName)
}
\arguments{
\item{rasterName}{filename (character) of the network raster}

\item{inName}{filename (character) of the point vector}

\item{outName}{filename (character) of the moved point vector}
}
\value{
A point vector file (\emph{outName}). Includes as attribute the
coordinates of the new locations and the Euclidean distance between the
original and moved location
}
\description{
Travel cost computations are performed between locations (nodes)
of a graph representing the road network. Points that are not on the network
must first move from their original location. The new location is determined
as the one on the network closest (in 2D) to the original location.
}
\details{
\emph{movePt2RdSegment()} must be run after \emph{createGraph()} to
ensure that any disconnected road network segments have been already removed
from \emph{rasterName}. Distribution moments of the move distances are shown
on the console
}
\examples{
\dontrun{
rdPath <- system.file( "data", "roads.gpkg", package="tRee2MillCost" )
fromPath <- system.file( "data", "sample.gpkg", package="tRee2MillCost" )
rasterizeRoads( rdPath, "MPH", 20, "road_net.tif" )
graph <- createGraph( "road_net.tif" )
movePt2RdSegment( "road_net.tif", fromPath, "movedFromLocations.gpkg" )
p <- terra::vect( "movedFromLocations.gpkg" )
head( data.frame(p) )
}

}
