% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRee2MillCost_functions_20260825.R
\name{setCellValueToInteger}
\alias{setCellValueToInteger}
\title{Convert a numeric raster to integer type}
\usage{
setCellValueToInteger(r, name)
}
\arguments{
\item{r}{spatRast object}

\item{name}{character, filename/path for the output integer raster}
}
\description{
Convert a numeric raster to integer type
}
\details{
The range of values in \emph{r} should in 0-255. The data type of
the output is set to INT1U. Large rasters will be processed in tiles. TIF
format is expected. If the file exists, it will be overwritten
}
